Added dynamic medication survey to mood poll; Started adding high contrast mode colors
